Moved ContextualToolbar to the `contextual` subdiredtory.
