+++ src/toolbar/contextual/contextualtoolbar.js
@@ -4,13 +4,13 @@
  */
 
 /**
- * @module ui/toolbar/contextualtoolbar
+ * @module ui/toolbar/contextual/contextualtoolbar
  */
 
 import Plugin from '@ckeditor/ckeditor5-core/src/plugin';
-import ContextualBalloon from '../contextualballoon';
-import ToolbarView from './toolbarview';
-import BalloonPanelView from '../panel/balloon/balloonpanelview.js';
+import ContextualBalloon from '../../contextualballoon';
+import ToolbarView from '../toolbarview';
+import BalloonPanelView from '../../panel/balloon/balloonpanelview.js';
 import debounce from '@ckeditor/ckeditor5-utils/src/lib/lodash/debounce';
 
 const defaultPositions = BalloonPanelView.defaultPositions;
